requirements.txt
aiomysql==0.2.0
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
packaging==24.2
pydantic==2.10.2
pydantic_core==2.27.1
PyMySQL==1.1.1
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
